Use lookup table for armor types to simplify armor generation
This centralizes the armor classifications into a hashtable-of-hashtables, where each
inner hashtable has all the various properties that were defined in different locations
and made it more onerous to change things.

This should clarify that each armor type has a certain number of elemental modifiers,
and what stat ranges should go with them.